refactor: rename q1-q4 folders to topic names and add README/TEST_LOG
