+++ weeks/week-18/solutions/1114405054/BinarySearch/README.md
@@ -1,0 +1,17 @@
+# 第四題：二分搜尋效能
+
+## 執行方式
+```bash
+python binary_search.py
+```
+
+## 題目規格
+1. 產生升冪排序的整數陣列
+2. 實作二分搜尋找出目標 K
+3. 輸出是否存在與比較次數
+4. 用 timeit 量測線性搜尋與二分搜尋耗時
+5. 畫雷達圖輸出至 `radar.png`
+
+## 個人參數
+- 學號末兩碼 = 54
+- K = 154
