+++ weeks/week-18/solutions/1114405054/BinarySearch/binary_search.py
