+++ weeks/week-18/solutions/1114405054/BinarySearch/test_binary_search.py
@@ -1,5 +1,5 @@
 import unittest
-from q4 import linear_search, binary_search
+from binary_search import linear_search, binary_search
 
 
 class TestSearch(unittest.TestCase):
