+++ weeks/week-18/solutions/1114405054/CaesarCipher/README.md
@@ -1,0 +1,16 @@
+# 第二題：凱撒密碼 (Caesar Cipher)
+
+## 執行方式
+```bash
+python caesar_cipher.py < input.txt
+```
+
+## 題目規格
+將每行字串中的英文字母向後位移 SHIFT 位：
+- 大寫在 `A-Z` 內循環
+- 小寫在 `a-z` 內循環
+- 非英文字母字元原樣保留
+
+## 個人參數
+- 學號末兩碼 = 54
+- SHIFT = 5
