+++ weeks/week-18/solutions/1114405054/CaesarCipher/test_caesar_cipher.py
@@ -1,5 +1,5 @@
 import unittest
-from q2 import caesar_cipher
+from caesar_cipher import caesar_cipher
 
 
 class TestCaesarCipher(unittest.TestCase):
