+++ weeks/week-18/solutions/1114405054/DataCleaning/README.md
@@ -1,0 +1,16 @@
+# 第一題：資料清理 (Data Cleaning)
+
+## 執行方式
+```bash
+python data_cleaning.py < input.txt
+```
+
+## 題目規格
+對每一組整數數列，依序完成三步：
+1. 去除重複（保留第一次出現的順序）
+2. 只保留能被 D 整除的數
+3. 由小到大排序
+
+## 個人參數
+- 學號末兩碼 = 54
+- D = 2
