+++ weeks/week-18/solutions/1114405054/DataCleaning/test_data_cleaning.py
@@ -1,5 +1,5 @@
 import unittest
-from q1 import clean_data
+from data_cleaning import clean_data
 
 
 class TestCleanData(unittest.TestCase):
