+++ weeks/week-18/solutions/1114405054/DigitalRoot/README.md
@@ -1,0 +1,16 @@
+# 第三題：任意進位的數字根
+
+## 執行方式
+```bash
+python digital_root.py < input.txt
+```
+
+## 題目規格
+1. 對每個輸入的十進位非負整數 x，先換算成 base 進位
+2. 將各位數字相加得新數
+3. 重複此步驟直到結果為個位數
+4. 以十進位輸出最終數字根
+
+## 個人參數
+- 學號末兩碼 = 54
+- base = 5
