+++ weeks/week-18/solutions/1114405054/DigitalRoot/digital_root.py
